+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * If selected frame has a local or upval `_ENV`, evals will use it instead of the global environment
 * Removed Event Check
 * Expand functions's upvals as children in Variables view
+* `__DebugAdapter.levelPath` is no longer required, as it can be filled automatically from `script.level`
+
 ## 1.1.2
 
 * Show best-guess arguments for `__index` and `__newindex` in stack trace